+++ app/models/research_area.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class ResearchArea < ApplicationRecord
-  has_ancestry cache_depth: true
+  include Parentable
 
   has_many :service_research_areas, autosave: true, dependent: :destroy
   has_many :services, through: :service_research_areas