+++ db/migrate/20190122205054_add_ancestry_depth_to_category.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddAncestryDepthToCategory < ActiveRecord::Migration[5.2]
+  def change
+    add_column :categories, :ancestry_depth, :integer, default: 0
+  end
+end