+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - User needs to select "Research area" while ordering service (@mkasztelnik)
 - Add voucher support in service ordering (@michal-szostak)
+- Basic CRUD for backoffice Research areas (@mkasztelnik)
+- Basic CRUD for backoffice Categories (@mkasztelnik)
 
 ### Changed